applied the ESH code-cleanup rules, update license header

+++ bundles/binding/org.openhab.binding.davis/src/main/java/org/openhab/binding/davis/datatypes/DataTypeWindHiRes.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2010-2015, openHAB.org and others.
+ * Copyright (c) 2010-2016, openHAB.org and others.
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -15,20 +15,20 @@ import org.openhab.core.library.types.DecimalType;
 import org.openhab.core.types.State;
 
 /**
- * Class to handle wind speeds 
+ * Class to handle wind speeds
  * 2 bytes, encoded in 0.1 mph, result in km/h
- * 
+ *
  * @author Trathnigg Thomas
  * @since 1.6.0
  */
 public class DataTypeWindHiRes implements DavisDataType {
 
-	/**
-	 * {@inheritDoc}
-	 */
-	public State convertToState(byte[] data, DavisValueType valueType) {
-		short value = ByteBuffer.wrap(data).order(ByteOrder.LITTLE_ENDIAN).getShort(valueType.getDataOffset());		
-		return new DecimalType((double)value*0.1 * 1.609344);
-	}
+    /**
+     * {@inheritDoc}
+     */
+    public State convertToState(byte[] data, DavisValueType valueType) {
+        short value = ByteBuffer.wrap(data).order(ByteOrder.LITTLE_ENDIAN).getShort(valueType.getDataOffset());
+        return new DecimalType(value * 0.1 * 1.609344);
+    }
 
 }
